+++ packages/formdata-event/ts_src/index.ts
@@ -10,8 +10,8 @@
  */
 
 import {install as installEvent} from "./Wrappers/Event.js";
-import {install as installFormDataEvent} from "./Wrappers/FormDataEvent.js";
 import {install as installEventTarget} from "./Wrappers/EventTarget.js";
+import {install as installFormDataEvent} from "./FormDataEvent.js";
 
 const priorFormDataEvent = window['FormDataEvent'] as any;
 if (priorFormDataEvent === undefined ||
